Integrate virt_wifi_sim driver in trout kernel
This driver commonly exists as a module for cuttlefish in a separate
repository, but because the trout kernel is running as a monolith within
the ACK repo, we need to bring the virt_wifi_sim module in and standalone
in the trout codebase.

Loading this driver restores the VirtWiFi wireless network for trout
devices, and allows connecting to it via a simulated wlan0 adapter.

Bug: 162251630
Test: connect to VirtWifi
Change-Id: I07f6eba6967c5fd0d17c5abdfe934c1e1c0fd664